Move the built-in SnConfigurationProvider class to a separate file.
