+++ src/SenseNet.Tools/SenseNet.Tools.csproj
@@ -50,6 +50,7 @@
   <ItemGroup>
     <Compile Include="Configuration\IConfigProvider.cs" />
     <Compile Include="Configuration\SnConfig.cs" />
+    <Compile Include="Configuration\SnConfigProvider.cs" />
     <Compile Include="Tools\CommandLineArguments\Argument.cs" />
     <Compile Include="Tools\CommandLineArguments\NoNameOptionAttribute.cs" />
     <Compile Include="Tools\CommandLineArguments\OptionAttribute.cs" />
